Reorganize module into multiple files
