+++ src/scripts/module.ts
@@ -1,18 +1,6 @@
-import * as uuid from "uuid";
-
-const FOUNDRY_ID_FLAG = "foundry_redirect_id"
-
-const SERVER_BASE_URL = "https://foundryredirect.com"
-const FOUNDRY_ID_URL_PARAM = "foundry_id";
-const EXTERNAL_ADDRESS_URL_PARAM = "external_address";
-const INTERNAL_ADDRESS_URL_PARAM = "internal_address";
-
-const PUBLIC_ID_KEY = "public_id";
-
-interface RedirectAddresses {
-    externalAddress : string,
-    localAddress : string
-}
+import { getOrCreateFoundryId, getUser } from "./foundryUtils";
+import { debugLog, displayErrorMessageToUser } from "./logging";
+import { getRedirectAddress, postFoundryInfo } from "./server";
 
 async function refreshIpData() : Promise<void> {
     debugLog("Foundry Redirect: Refreshing foundry link data");
@@ -39,70 +27,6 @@ async function refreshIpData() : Promise<void> {
 
     setTimeout(refreshIpData, 1000 * 60 * 60)
     return p;
-}
-
-function getOrCreateFoundryId() : string {
-    let user = getUser();
-    let foundryId = user?.getFlag("core", FOUNDRY_ID_FLAG);
-    if(!foundryId){
-        console.log("No foundry redirect ID found. Generating one...")
-        foundryId = uuid.v1();
-        user?.setFlag("core", FOUNDRY_ID_FLAG, foundryId);
-    }
-    debugLog("Foundry Redirect Id: " + foundryId)
-    return <string>foundryId;
-}
-
-
-async function postFoundryInfo(foundryId : string, externalAddress:string, localAddress: string) : Promise<void> {
-    return fetch(`${SERVER_BASE_URL}?${FOUNDRY_ID_URL_PARAM}=${foundryId}&${EXTERNAL_ADDRESS_URL_PARAM}=${externalAddress}&${INTERNAL_ADDRESS_URL_PARAM}=${localAddress}`, {
-        method: "POST"
-    }).then(res=>{
-        debugLog("Foundry redirect: Successfully updated server address on server")
-    }).catch(err=>{
-        displayErrorMessageToUser("Failed to post server address to redirect server")
-        console.error(err);
-    })
-}
-
-async function getRedirectAddress() : Promise<RedirectAddresses|undefined> { 
-    const foundryId = getOrCreateFoundryId();
-    return fetch(`${SERVER_BASE_URL}?${FOUNDRY_ID_URL_PARAM}=${foundryId}`).then(async res =>{
-        let responseText  = await res.text();
-        debugLog("Fetch redirect address response: " + responseText);
-        let redirect : RedirectAddresses = {
-            externalAddress: responseText,
-            localAddress : responseText + "/local"
-        }
-        return redirect;
-    }).catch(err=>{
-        displayErrorMessageToUser("Failed to fetch foundry redirect address from server")
-        console.error(err)
-        return undefined;
-    });
-}
-
-function displayInfoMessageToUser(message:string) {
-    message = "Foundry Redirect: " + message;
-    ui.notifications?.info(message);
-    console.log(message);
-}
-
-function displayErrorMessageToUser(error:string) {
-    error = "Foundry Redirect: " + error
-    ui.notifications?.error(error);
-    console.error(error);
-}
-
-function getUser() : StoredDocument<User> | null {
-    let g = <Game>game
-    return g.user;
-}
-
-function debugLog(msg:any) {
-    if( CONFIG.debug.hasOwnProperty("redirect") && CONFIG.debug["redirect"]){
-        console.log(msg);
-    }
 }
   
 Hooks.on("ready", function() {
